refactor(storage): split sqlite.rs into module dir (bo-q19x)
src/storage/sqlite.rs had grown to 1011 lines (over the 500-line soft-limit,
not allowlisted) after the bead_lineage/bug_causality extensions. Split into a
cohesive module directory with no behavior change — methods and structs are
moved verbatim:

- sqlite/mod.rs (326): MetadataStore + core (open/schema/migrate/txn), meta,
  coupling, and file-hash methods; submodule decls; re-exports the lineage /
  causality structs so external paths (storage::sqlite::{NewBeadLineage, ...})
  are unchanged.
- sqlite/lineage.rs (297): bead-lineage + bug-causality impl methods and their
  structs (PriorTouch, New/Record types, TouchedSymbol).
- sqlite/tests.rs (398): the existing #[cfg(test)] suite.

All three files are under the 500-line soft-limit. Verified: cargo build clean,
cargo clippy clean (no new warnings on the split), full suite 862 passed / 0
failed (sqlite module: 16/16). No fmt churn — bobbin CI does not gate fmt and
main is not fmt-clean, so only these files changed.
